+++ services/src/main/java/com/codigoartesanal/entuliga/services/impl/JugadorServiceImpl.java
@@ -7,6 +7,7 @@ import com.codigoartesanal.entuliga.services.JugadorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.*;
 
 /**
@@ -117,11 +118,11 @@ public class JugadorServiceImpl implements JugadorService {
         geoLocation.setCodigoPostal(ligaMap.get(PROPERTY_CODIGO_POSTAL));
         String latitude = ligaMap.get(PROPERTY_LATITUDE);
         if (latitude != null && !latitude.isEmpty()) {
-            geoLocation.setLatitude(Double.parseDouble(latitude));
+            geoLocation.setLatitude(BigDecimal.valueOf(Double.parseDouble(latitude)));
         }
         String longitude = ligaMap.get(PROPERTY_LONGITUDE);
         if (longitude != null && !longitude.isEmpty()) {
-            geoLocation.setLongitude(Double.parseDouble(longitude));
+            geoLocation.setLongitude(BigDecimal.valueOf(Double.parseDouble(longitude)));
         }
         Colonia colonia = new Colonia();
         colonia.setId(Long.valueOf(ligaMap.get(PROPERTY_COLONIA_ID)));
